LPS-170489 Rename ModuleApplicationContextExtender for ModuleApplicationContextExtension. Register LiferayPortalServiceExtension service when bundle matches with its condition, then LiferayPortalServiceExtensionServiceTracker tracks it to start ModuleApplicationContextExtension.
